requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
bcrypt>=3.2.0
pyotp>=2.6.0
fpdf2>=2.5.5
python-dotenv>=0.19.0